Add files via upload
Magic Number is interesting project developed in C++, in this the user has to think of number between the given range. Then the user will be provided with number sets, he has to give answer in yes/no form. At the end, the game will display the number user thought of.